good job

+++ app/Commands/Install.php
@@ -61,8 +61,23 @@ class Install extends Command
         // shell_exec('unzip '.$local_file.' -d '.$path);
         shell_exec(' curl https://phpyolk.com/Archive.zip --progress-bar  --output '.$projectname.'.zip');
         $this->info('Unzipping........');
-        shell_exec('unzip '.$projectname.'.zip -d '.$path.'/'.$projectname);
-        shell_exec('rm '.$projectname.'.zip');
+        $this->info('Please wait...');
+
+
+        if (strtoupper(substr(PHP_OS, 0, 3)) === 'WIN') {
+                if(!is_dir($path.'/'.$projectname))
+                {
+                    mkdir($path.'/'.$projectname, 0777);
+                }
+            $this->info('Copying files........');
+            shell_exec('tar -zxvf' .$projectname.'.zip -C' .$path .'/'.$projectname); 
+            $this->info('setting up yolk environment........');
+            shell_exec( 'Del '.$projectname.'.zip');
+        } else {
+            shell_exec('unzip '.$projectname.'.zip -d '.$path.'/'.$projectname);
+             shell_exec('rm '.$projectname.'.zip');
+        }
+       
 
         $this->info('Done!');
         $this->info('Your new app is ready to go!');
